+++ src/main/java/com/phion/spring/beans/BeanDefinitionReader.java
@@ -1,4 +1,4 @@
-package com.phion.spring.ioc.beans;
+package com.phion.spring.beans;
 
 /**
  * bean加载器定义