+++ src/main/java/com/phion/spring/beans/factory/AbstractBeanFactory.java
@@ -1,6 +1,6 @@
-package com.phion.spring.ioc.beans.factory;
+package com.phion.spring.beans.factory;
 
-import com.phion.spring.ioc.beans.BeanDefinition;
+import com.phion.spring.beans.BeanDefinition;
 import lombok.extern.slf4j.Slf4j;
 
 import java.util.*;
@@ -8,7 +8,7 @@ import java.util.concurrent.ConcurrentHashMap;
 
 /**
  * 抽象bean工厂，管理所有的bean
- *
+ * <p>
  * 提供默认的注册与获取bean的方法
  *
  * @author yanful
@@ -26,33 +26,34 @@ public abstract class AbstractBeanFactory implements BeanFactory {
 
     /**
      * 对延迟加载适配
+     *
      * @param name bean名称
      * @return bean
      */
     @Override
     public Object getBean(String name) throws Exception {
         BeanDefinition beanDefinition = beanDefinitionMap.get(name);
-        if(Objects.isNull(beanDefinition)){
+        if (Objects.isNull(beanDefinition)) {
             throw new IllegalArgumentException("No bean named " + name + " is defined");
         }
         Object bean = beanDefinition.getBean();
-        if(Objects.isNull(bean)){
-            bean =doCreateBean(beanDefinition);
+        if (Objects.isNull(bean)) {
+            bean = doCreateBean(beanDefinition);
         }
         return bean;
     }
 
     /**
      * 注入对象的方法
-     *
+     * <p>
      * 默认使用延迟加载，即使用时才实例化bean
      *
-     * @param name 对象名称，唯一标识
+     * @param name           对象名称，唯一标识
      * @param beanDefinition 对象定义信息
      */
-    public void registerBeanDefinition(String name,BeanDefinition beanDefinition) throws Exception {
+    public void registerBeanDefinition(String name, BeanDefinition beanDefinition) throws Exception {
 
-        beanDefinitionMap.put(name,beanDefinition);
+        beanDefinitionMap.put(name, beanDefinition);
         beanDefinitionNames.add(name);
     }
 
@@ -67,6 +68,7 @@ public abstract class AbstractBeanFactory implements BeanFactory {
 
     /**
      * 通过类定义信息注入默认的bean，抽象方法，给子类实现
+     *
      * @param beanDefinition bean定义信息
      * @return 注入的bean
      */