+++ src/main/java/com/phion/spring/beans/xml/XmlBeanDefinitionReader.java
@@ -1,11 +1,11 @@
-package com.phion.spring.ioc.beans.xml;
+package com.phion.spring.beans.xml;
 
-import com.phion.spring.ioc.*;
-import com.phion.spring.ioc.beans.AbstractBeanDefinitionReader;
-import com.phion.spring.ioc.beans.BeanDefinition;
-import com.phion.spring.ioc.beans.PropertyValue;
-import com.phion.spring.ioc.beans.PropertyValues;
-import com.phion.spring.ioc.beans.io.ResourceLoader;
+import com.phion.spring.BeanReference;
+import com.phion.spring.beans.AbstractBeanDefinitionReader;
+import com.phion.spring.beans.BeanDefinition;
+import com.phion.spring.beans.PropertyValue;
+import com.phion.spring.beans.PropertyValues;
+import com.phion.spring.beans.io.ResourceLoader;
 import com.phion.spring.support.CollectionsUtils;
 import com.phion.spring.support.StringUtils;
 import lombok.extern.slf4j.Slf4j;